requirements.txt
# Core async framework
fastapi
uvicorn[standard]
pydantic

# Data persistence
elasticsearch[async]
aiofiles

# Data science / ML
pandas
numpy
scikit-learn

# HTTP client (used by LLM integrations and Splunk sender)
aiohttp

# NLP / log sentiment analysis
vaderSentiment

# Configuration
python-dotenv

# Observability dashboard
streamlit
altair

# Testing
pytest
pytest-asyncio

# Documentation
mkdocs
mkdocs-material
mkdocs-include-markdown-plugin